split chess-analyzer.tsx into per-tab chart modules
extracted five chart renderers, four tooltips, and the overview cards
out of the ~1070-line god component into components/charts/ and
components/overview-cards.tsx:

- chart-tooltips.tsx: CustomTooltip, CustomMonthlyTooltip, RatingTooltip,
  renderCustomizedTick
- themed-brush.tsx: themedBrushProps(theme) helper (dedupes the
  4x-repeated theme-dependent Brush styling block)
- monthly-performance-chart.tsx, openings-chart.tsx,
  rating-progression-chart.tsx, head-to-head-chart.tsx
- overview-cards.tsx: the five overview-tab MagicCard blocks

Pure extraction, no behavior change: same props, colors, brush indices.
The rating chart now uses RatingTooltip instead of an inline duplicate
(the plan's own note confirmed they render identical fields).

chess-analyzer.tsx drops from ~1070 to 580 lines and is now completely
lint-clean (zero no-unused-vars warnings, including the RatingTooltip
warning that existed before this plan). Verified with a live headless
browser: uploaded a PGN, clicked all 5 tabs, toggled dark/light theme,
opened the share dialog — zero console errors, brush colors and stats
correct in both themes (screenshots checked).

Note: file is 580 lines, not under the plan's <500 target. The
remaining bulk is the input-card JSX + help-tooltip logic, which the
plan's own "final shape" description says should stay in this file;
splitting that further wasn't in the plan's authorized file list, so
left as-is rather than inventing an unscoped extraction.
